+++ sidebars.js
@@ -411,7 +411,7 @@ const sidebars = {
             },
             {
               type: "category",
-              label: "Use Bot APIs",
+              label: "Bot APIs",
               items: [
                 "common-workflows/analytics/use-bot-api/use-bot-api",
                 "common-workflows/analytics/use-bot-api/get-bot-configuration",
@@ -421,6 +421,17 @@ const sidebars = {
                 "common-workflows/analytics/use-bot-api/ask-bot-instance-question",
                 "common-workflows/analytics/use-bot-api/get-a-bot-question",
                 "common-workflows/analytics/use-bot-api/delete-a-bot-instance",
+              ],
+            },
+            {
+              type: "category",
+              label: "Question with Bot Routing APIs",
+              items: [
+                "common-workflows/analytics/question-with-bot-routing-api/question-api",
+                "common-workflows/analytics/question-with-bot-routing-api/ask-question-to-bots-in-the-application",
+                "common-workflows/analytics/question-with-bot-routing-api/get-question-from-bots-in-the-application",
+                "common-workflows/analytics/question-with-bot-routing-api/get-image-of-answer",
+                "common-workflows/analytics/question-with-bot-routing-api/ask-suggested-question-to-bots-in-the-application",
               ],
             },
           ],